rename tabList to tab-list
